unify loading of missing data
